cleanup infra/domain
