+++ server/src/domain/metadata/metadata.repository.ts
@@ -1,5 +1,4 @@
-import { AssetEntity, ExifEntity } from '@app/infra/entities';
-import { ReadTaskOptions, Tags } from 'exiftool-vendored';
+import { Tags } from 'exiftool-vendored';
 import { InitOptions } from 'local-reverse-geocoder';
 
 export const IMetadataRepository = 'IMetadataRepository';
@@ -28,8 +27,5 @@ export interface IMetadataRepository {
   init(options: Partial<InitOptions>): Promise<void>;
   reverseGeocode(point: GeoPoint): Promise<ReverseGeocodeResult>;
   deleteCache(): Promise<void>;
-  getDuration(seconds?: number): string;
-  getExifTags(path: string, options?: Partial<ReadTaskOptions>): Promise<ImmichTags | null>;
-  getTimezone(lat: number, lon: number): string;
-  mapExifEntity(asset: AssetEntity, tags: ImmichTags, fileSize: number): ExifEntity;
+  getExifTags(path: string): Promise<ImmichTags | null>;
 }
