+++ server/src/domain/storage/storage.repository.ts
@@ -1,5 +1,5 @@
 import { Stats } from 'fs';
-import fs from 'fs/promises';
+import { FileReadOptions } from 'fs/promises';
 import { Readable } from 'stream';
 import { CrawlOptionsDto } from '../library';
 
@@ -25,7 +25,7 @@ export const IStorageRepository = 'IStorageRepository';
 export interface IStorageRepository {
   createZipStream(): ImmichZipStream;
   createReadStream(filepath: string, mimeType?: string | null): Promise<ImmichReadStream>;
-  readFile(filepath: string, options?: fs.FileReadOptions<Buffer>): Promise<fs.FileReadResult<Buffer>>;
+  readFile(filepath: string, options?: FileReadOptions<Buffer>): Promise<Buffer>;
   writeFile(filepath: string, buffer: Buffer): Promise<void>;
   unlink(filepath: string): Promise<void>;
   unlinkDir(folder: string, options?: { recursive?: boolean; force?: boolean }): Promise<void>;
