+++ server/src/infra/repositories/filesystem.provider.ts
@@ -8,7 +8,7 @@ import {
 } from '@app/domain';
 import archiver from 'archiver';
 import { constants, createReadStream, existsSync, mkdirSync } from 'fs';
-import fs, { readdir } from 'fs/promises';
+import fs, { readdir, writeFile } from 'fs/promises';
 import { glob } from 'glob';
 import mv from 'mv';
 import { promisify } from 'node:util';
@@ -39,16 +39,17 @@ export class FilesystemProvider implements IStorageRepository {
     };
   }
 
-  async readFile(filepath: string, options?: fs.FileReadOptions<Buffer>): Promise<fs.FileReadResult<Buffer>> {
+  async readFile(filepath: string, options?: fs.FileReadOptions<Buffer>): Promise<Buffer> {
     const file = await fs.open(filepath);
-    const buffer = await file.read(options);
-    await file.close();
-    return buffer;
+    try {
+      const { buffer } = await file.read(options);
+      return buffer;
+    } finally {
+      await file.close();
+    }
   }
 
-  writeFile(filepath: string, buffer: Buffer): Promise<void> {
-    return fs.writeFile(filepath, buffer);
-  }
+  writeFile = writeFile;
 
   async moveFile(source: string, destination: string): Promise<void> {
     if (await this.checkFileExists(destination)) {
