+++ server/test/repositories/metadata.repository.mock.ts
@@ -3,11 +3,8 @@ import { IMetadataRepository } from '@app/domain';
 export const newMetadataRepositoryMock = (): jest.Mocked<IMetadataRepository> => {
   return {
     deleteCache: jest.fn(),
-    getDuration: jest.fn(),
     getExifTags: jest.fn(),
-    getTimezone: jest.fn(),
     init: jest.fn(),
-    mapExifEntity: jest.fn(),
     reverseGeocode: jest.fn(),
   };
 };
